requirements.txt
requests
docopt
selenium-requests
